+++ tests/test_decompressor.py
@@ -487,7 +487,7 @@ class TestDecompressor_stream_reader(unittest.TestCase):
         self.assertFalse(buffer.closed)
         reader.close()
         self.assertTrue(reader.closed)
-        self.assertFalse(buffer.closed)
+        self.assertTrue(buffer.closed)
 
         with self.assertRaisesRegex(ValueError, "stream is closed"):
             reader.read(b"")
@@ -504,7 +504,7 @@ class TestDecompressor_stream_reader(unittest.TestCase):
             reader.read(3)
 
         self.assertTrue(reader.closed)
-        self.assertFalse(buffer.closed)
+        self.assertTrue(buffer.closed)
 
         # Context manager exit should close stream if an exception raised.
         buffer = io.BytesIO(foo)
@@ -516,7 +516,7 @@ class TestDecompressor_stream_reader(unittest.TestCase):
                 raise Exception("ignore")
 
         self.assertTrue(reader.closed)
-        self.assertFalse(buffer.closed)
+        self.assertTrue(buffer.closed)
 
         # Test with non-file source variant.
         with dctx.stream_reader(foo) as reader:
@@ -525,19 +525,19 @@ class TestDecompressor_stream_reader(unittest.TestCase):
 
         self.assertTrue(reader.closed)
 
-    def test_close_closefd_true(self):
+    def test_close_closefd_false(self):
         foo = zstd.ZstdCompressor().compress(b"foo" * 1024)
 
         buffer = io.BytesIO(foo)
         dctx = zstd.ZstdDecompressor()
-        reader = dctx.stream_reader(buffer, closefd=True)
+        reader = dctx.stream_reader(buffer, closefd=False)
 
         reader.read(3)
         self.assertFalse(reader.closed)
         self.assertFalse(buffer.closed)
         reader.close()
         self.assertTrue(reader.closed)
-        self.assertTrue(buffer.closed)
+        self.assertFalse(buffer.closed)
 
         with self.assertRaisesRegex(ValueError, "stream is closed"):
             reader.read(b"")
@@ -548,17 +548,17 @@ class TestDecompressor_stream_reader(unittest.TestCase):
 
         # Context manager exit should not close stream.
         buffer = io.BytesIO(foo)
-        reader = dctx.stream_reader(buffer, closefd=True)
+        reader = dctx.stream_reader(buffer, closefd=False)
 
         with reader:
             reader.read(3)
 
         self.assertTrue(reader.closed)
-        self.assertTrue(buffer.closed)
+        self.assertFalse(buffer.closed)
 
         # Context manager exit should close stream if an exception raised.
         buffer = io.BytesIO(foo)
-        reader = dctx.stream_reader(buffer, closefd=True)
+        reader = dctx.stream_reader(buffer, closefd=False)
 
         with self.assertRaisesRegex(Exception, "ignore"):
             with reader:
@@ -566,10 +566,10 @@ class TestDecompressor_stream_reader(unittest.TestCase):
                 raise Exception("ignore")
 
         self.assertTrue(reader.closed)
-        self.assertTrue(buffer.closed)
+        self.assertFalse(buffer.closed)
 
         # Test with non-file source variant.
-        with dctx.stream_reader(foo, closefd=True) as reader:
+        with dctx.stream_reader(foo, closefd=False) as reader:
             reader.read(3)
             self.assertFalse(reader.closed)